[fix]: 구조 변경(#1)
