+++ src/main/java/com/sinabro/server/domain/repository/PlacesRepository.java
@@ -1,5 +1,6 @@
-package com.sinabro.server.model.places;
+package com.sinabro.server.domain.repository;
 
+import com.sinabro.server.domain.entity.Places;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
